Chore: remove duplicated SubjetInfo code in tests.
