+++ src/simulator/equation/analytical/one_compartment_cl_models.rs
@@ -41,63 +41,10 @@ pub fn one_compartment_cl_with_absorption(x: &V, p: &V, t: T, rateiv: V, cov: &C
 
 #[cfg(test)]
 mod tests {
+    use super::super::tests::SubjectInfo;
     use super::{one_compartment_cl, one_compartment_cl_with_absorption};
     use crate::*;
     use approx::assert_relative_eq;
-
-    enum SubjectInfo {
-        InfusionDosing,
-        OralInfusionDosage,
-    }
-
-    impl SubjectInfo {
-        fn get_subject(&self) -> Subject {
-            match self {
-                SubjectInfo::InfusionDosing => Subject::builder("id1")
-                    .bolus(0.0, 100.0, 0)
-                    .infusion(24.0, 150.0, 0, 3.0)
-                    .missing_observation(0.0, 0)
-                    .missing_observation(1.0, 0)
-                    .missing_observation(2.0, 0)
-                    .missing_observation(4.0, 0)
-                    .missing_observation(8.0, 0)
-                    .missing_observation(12.0, 0)
-                    .missing_observation(24.0, 0)
-                    .missing_observation(25.0, 0)
-                    .missing_observation(26.0, 0)
-                    .missing_observation(27.0, 0)
-                    .missing_observation(28.0, 0)
-                    .missing_observation(32.0, 0)
-                    .missing_observation(36.0, 0)
-                    .build(),
-
-                SubjectInfo::OralInfusionDosage => Subject::builder("id1")
-                    .bolus(0.0, 100.0, 1)
-                    .infusion(24.0, 150.0, 0, 3.0)
-                    .bolus(48.0, 100.0, 0)
-                    .missing_observation(0.0, 0)
-                    .missing_observation(1.0, 0)
-                    .missing_observation(2.0, 0)
-                    .missing_observation(4.0, 0)
-                    .missing_observation(8.0, 0)
-                    .missing_observation(12.0, 0)
-                    .missing_observation(24.0, 0)
-                    .missing_observation(25.0, 0)
-                    .missing_observation(26.0, 0)
-                    .missing_observation(27.0, 0)
-                    .missing_observation(28.0, 0)
-                    .missing_observation(32.0, 0)
-                    .missing_observation(36.0, 0)
-                    .missing_observation(48.0, 0)
-                    .missing_observation(49.0, 0)
-                    .missing_observation(50.0, 0)
-                    .missing_observation(52.0, 0)
-                    .missing_observation(56.0, 0)
-                    .missing_observation(60.0, 0)
-                    .build(),
-            }
-        }
-    }
 
     #[test]
     fn test_one_compartment_cl() {
